requirements.txt
gitpython 
langchain
chromadb
tiktoken
google.generativeai
markdown2
flask
flask-cors